Feat: Dynamic button size in CustomKeyboard
Resolves #35
