+++ KBoard/CustomKeyboard/KeyboardViewController.swift
@@ -25,7 +25,7 @@ class KeyboardViewController: UIInputViewController {
         view.showsVerticalScrollIndicator = true
         view.scrollIndicatorInsets = UIEdgeInsets(top: 2, left: 0, bottom: 0, right: 4)
         view.contentInset = .zero
-        view.backgroundColor = .clear
+        view.backgroundColor = .blue
         view.clipsToBounds = true
 //        view.register(MyCell.self, forCellWithReuseIdentifier: MyCell.id)
         view.translatesAutoresizingMaskIntoConstraints = false
@@ -53,7 +53,8 @@ class KeyboardViewController: UIInputViewController {
         self.nextKeyboardButton.leftAnchor.constraint(equalTo: self.view.leftAnchor).isActive = true
         self.nextKeyboardButton.bottomAnchor.constraint(equalTo: self.view.bottomAnchor).isActive = true
 
-        self.collectionView.delegate = self
+//        self.collectionView.delegate = self
+//        self.collectionView.dataSource = self
 
         let row1 = ["보라해💜", "💜JungKook💜", "💜JIN💜"]
         let row2 = ["최애", "존멋", "💜제이홉💜", "LOVE"]
@@ -64,11 +65,17 @@ class KeyboardViewController: UIInputViewController {
         var list2 = createRow(buttonTitles: row2)
         var list3 = createRow(buttonTitles: row3)
         var list4 = createRow(buttonTitles: row4)
-
+        var list: [UIView] = [list1, list2, list3, list4]
         self.view.addSubview(list1)
         self.view.addSubview(list2)
         self.view.addSubview(list3)
         self.view.addSubview(list4)
+//        self.view.addSubview(collectionView)
+
+//        self.view.addSubview(list1)
+//        self.view.addSubview(list2)
+//        self.view.addSubview(list3)
+//        self.view.addSubview(list4)
 
         list1.translatesAutoresizingMaskIntoConstraints = false
         list2.translatesAutoresizingMaskIntoConstraints = false
@@ -78,10 +85,15 @@ class KeyboardViewController: UIInputViewController {
         rowConstraints(view: self.view, rowViews: [list1, list2, list3, list4])
     }
 
-    func createButton(name: String) -> UIButton {
+    private func textSize(text: String) -> CGFloat {
+        return (text as NSString).size(withAttributes: [NSAttributedString.Key.font: UIFont(name: "Helvetica", size: 17)]).width
+    }
+
+    private func createButton(name: String) -> UIButton {
         let button = UIButton()
         button.translatesAutoresizingMaskIntoConstraints = false
         button.heightAnchor.constraint(equalToConstant: 40).isActive = true
+        button.widthAnchor.constraint(equalToConstant: textSize(text: name) + 40).isActive = true
         button.layer.cornerRadius = 20
         button.layer.borderColor = UIColor.black.cgColor
 //        button.layer.masksToBounds = true
@@ -98,9 +110,9 @@ class KeyboardViewController: UIInputViewController {
         proxy.insertText(text!)
     }
 
-    func createRow(buttonTitles: [String]) -> UIView {
+    private func createRow(buttonTitles: [String]) -> UIView {
         var buttons: [UIButton] = []
-        var row = UIView(frame: CGRect.init(x: 0, y: 0, width: 320, height: 10))
+        var row = UIView(frame: CGRect.init(x: 0, y: 0, width: 320, height: 0))
         row.backgroundColor = UIColor.systemGray6
         for buttonTitle in buttonTitles {
             let button = createButton(name: buttonTitle)
@@ -111,39 +123,39 @@ class KeyboardViewController: UIInputViewController {
         return row
     }
 
-    func buttonConstraints(buttons: [UIButton], view: UIView) {
+    private func buttonConstraints(buttons: [UIButton], view: UIView) {
         for (index, button) in buttons.enumerated() {
             var topConstraint = NSLayoutConstraint(item: button, attribute: .top, relatedBy: .equal, toItem: view, attribute: .top, multiplier: 1.0, constant: 12)
             var bottomConstraint = NSLayoutConstraint(item: button, attribute: .bottom, relatedBy: .equal, toItem: view, attribute: .bottom, multiplier: 1.0, constant: -3)
-            var rightConstraint: NSLayoutConstraint!
-            if index == buttons.count - 1 {
-                rightConstraint = NSLayoutConstraint(item: button, attribute: .right, relatedBy: .equal, toItem: view, attribute: .right, multiplier: 1.0, constant: -10)
-            } else {
-                let nextButton = buttons[index+1]
-                rightConstraint = NSLayoutConstraint(item: button, attribute: .right, relatedBy: .equal, toItem: nextButton, attribute: .left, multiplier: 1.0, constant: -10)
-            }
+//            var rightConstraint: NSLayoutConstraint!
+//            if index == buttons.count - 1 {
+//                rightConstraint = NSLayoutConstraint(item: button, attribute: .right, relatedBy: .equal, toItem: view, attribute: .right, multiplier: 1.0, constant: -10)
+//            } else {
+//                let nextButton = buttons[index+1]
+//                rightConstraint = NSLayoutConstraint(item: button, attribute: .right, relatedBy: .equal, toItem: nextButton, attribute: .left, multiplier: 1.0, constant: -10)
+//            }
             var leftConstraint: NSLayoutConstraint!
             if index == 0 {
                 leftConstraint = NSLayoutConstraint(item: button, attribute: .left, relatedBy: .equal, toItem: view, attribute: .left, multiplier: 1.0, constant: 12)
             } else {
                 let prevtButton = buttons[index-1]
                 leftConstraint = NSLayoutConstraint(item: button, attribute: .left, relatedBy: .equal, toItem: prevtButton, attribute: .right, multiplier: 1.0, constant: 10)
-                    let firstButton = buttons[0]
-                    var widthConstraint = NSLayoutConstraint(item: firstButton, attribute: .width, relatedBy: .equal, toItem: button, attribute: .width, multiplier: 1.0, constant: 0)
-                view.addConstraint(widthConstraint)
+                let firstButton = buttons[0]
+                var widthConstraint = NSLayoutConstraint(item: firstButton, attribute: .width, relatedBy: .equal, toItem: button, attribute: .width, multiplier: 1.0, constant: 0)
+//                view.addConstraint(widthConstraint)
             }
-            view.addConstraints([topConstraint, bottomConstraint, rightConstraint, leftConstraint])
+            view.addConstraints([topConstraint, bottomConstraint, leftConstraint])
         }
     }
 
-    func rowConstraints(view: UIView, rowViews: [UIView]) {
+    private func rowConstraints(view: UIView, rowViews: [UIView]) {
         for (index, rowView) in rowViews.enumerated() {
-            var rightSideConstraint = NSLayoutConstraint(item: rowView, attribute: .right, relatedBy: .equal, toItem: inputView, attribute: .right, multiplier: 1.0, constant: -1)
-            var leftConstraint = NSLayoutConstraint(item: rowView, attribute: .left, relatedBy: .equal, toItem: inputView, attribute: .left, multiplier: 1.0, constant: 1)
+            var rightSideConstraint = NSLayoutConstraint(item: rowView, attribute: .right, relatedBy: .equal, toItem: view, attribute: .right, multiplier: 1.0, constant: -1)
+            var leftConstraint = NSLayoutConstraint(item: rowView, attribute: .left, relatedBy: .equal, toItem: view, attribute: .left, multiplier: 1.0, constant: 1)
             view.addConstraints([leftConstraint, rightSideConstraint])
             var topConstraint: NSLayoutConstraint
             if index == 0 {
-                topConstraint = NSLayoutConstraint(item: rowView, attribute: .top, relatedBy: .equal, toItem: inputView, attribute: .top, multiplier: 1.0, constant: 0)
+                topConstraint = NSLayoutConstraint(item: rowView, attribute: .top, relatedBy: .equal, toItem: view, attribute: .top, multiplier: 1.0, constant: 0)
             } else {
                 let prevRow = rowViews[index-1]
                 topConstraint = NSLayoutConstraint(item: rowView, attribute: .top, relatedBy: .equal, toItem: prevRow, attribute: .bottom, multiplier: 1.0, constant: 0)
@@ -154,7 +166,7 @@ class KeyboardViewController: UIInputViewController {
             view.addConstraint(topConstraint)
             var bottomConstraint: NSLayoutConstraint
             if index == rowViews.count - 1 {
-                bottomConstraint = NSLayoutConstraint(item: rowView, attribute: .bottom, relatedBy: .equal, toItem: inputView, attribute: .bottom, multiplier: 1.0, constant: 0)
+                bottomConstraint = NSLayoutConstraint(item: rowView, attribute: .bottom, relatedBy: .equal, toItem: view, attribute: .bottom, multiplier: 1.0, constant: 0)
             } else {
                 let nextRow = rowViews[index+1]
                 bottomConstraint = NSLayoutConstraint(item: rowView, attribute: .bottom, relatedBy: .equal, toItem: nextRow, attribute: .top, multiplier: 1.0, constant: 0)
@@ -183,7 +195,7 @@ class KeyboardViewController: UIInputViewController {
     }
 }
 
-extension KeyboardViewController: UICollectionViewDelegateFlowLayout {
+ extension KeyboardViewController: UICollectionViewDelegateFlowLayout {
 
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
         return 1
@@ -195,7 +207,7 @@ extension KeyboardViewController: UICollectionViewDelegateFlowLayout {
 
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
 
-        let width = collectionView.frame.width / 3 - 1 ///  3등분하여 배치, 옆 간격이 1이므로 1을 빼줌
+        let width = collectionView.frame.width
         print("collectionView width=\(collectionView.frame.width)")
         print("cell하나당 width=\(width)")
         print("root view width = \(self.view.frame.width)")
@@ -203,77 +215,4 @@ extension KeyboardViewController: UICollectionViewDelegateFlowLayout {
         let size = CGSize(width: width, height: width)
         return size
     }
-}
-
-extension UIView {
-    func anchor(top: NSLayoutYAxisAnchor? = nil,
-                left: NSLayoutXAxisAnchor? = nil,
-                bottom: NSLayoutYAxisAnchor? = nil,
-                right: NSLayoutXAxisAnchor? = nil,
-                paddingTop: CGFloat = 0,
-                paddingLeft: CGFloat = 0,
-                paddingBottom: CGFloat = 0,
-                paddingRight: CGFloat = 0,
-                width: CGFloat? = nil,
-                height: CGFloat? = nil) {
-
-        translatesAutoresizingMaskIntoConstraints = false
-
-        if let top = top {
-            topAnchor.constraint(equalTo: top, constant: paddingTop).isActive = true
-        }
-
-        if let left = left {
-            leftAnchor.constraint(equalTo: left, constant: paddingLeft).isActive = true
-        }
-
-        if let bottom = bottom {
-            bottomAnchor.constraint(equalTo: bottom, constant: -paddingBottom).isActive = true
-        }
-
-        if let right = right {
-            rightAnchor.constraint(equalTo: right, constant: -paddingRight).isActive = true
-        }
-
-        if let width = width {
-            widthAnchor.constraint(equalToConstant: width).isActive = true
-        }
-
-        if let height = height {
-            heightAnchor.constraint(equalToConstant: height).isActive = true
-        }
-    }
-
-    func centerX(inView view: UIView) {
-        translatesAutoresizingMaskIntoConstraints = false
-        centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
-    }
-
-    func centerY(inView view: UIView, leftAnchor: NSLayoutXAxisAnchor? = nil,
-                 paddingLeft: CGFloat = 0, constant: CGFloat = 0) {
-
-        translatesAutoresizingMaskIntoConstraints = false
-        centerYAnchor.constraint(equalTo: view.centerYAnchor, constant: constant).isActive = true
-
-        if let left = leftAnchor {
-            anchor(left: left, paddingLeft: paddingLeft)
-        }
-    }
-
-    func setDimensions(height: CGFloat, width: CGFloat) {
-        translatesAutoresizingMaskIntoConstraints = false
-        heightAnchor.constraint(equalToConstant: height).isActive = true
-        widthAnchor.constraint(equalToConstant: width).isActive = true
-    }
-
-    func setHeight(height: CGFloat) {
-        translatesAutoresizingMaskIntoConstraints = false
-        heightAnchor.constraint(equalToConstant: height).isActive = true
-    }
-
-    func setWidth(width: CGFloat) {
-        translatesAutoresizingMaskIntoConstraints = false
-        widthAnchor.constraint(equalToConstant: width).isActive = true
-    }
-
-}
+ }
